+++ BALSAMIC/constants/workflow_rules.py
@@ -136,5 +136,5 @@ DELIVERY_RULES = [
     "mergeBam_normal",
     "cnvkit_paired",
     "cnvkit_single",
-    "ascat_tumor_normal",
+    "ascat_tumor_normal_merge_output",
 ]