+++ CHANGELOG.rst
@@ -4,7 +4,8 @@
 Added:
 ^^^^^^
 
-* Added ``--hgsvg`` annotation to VEP
+* ``--hgsvg`` annotation to VEP #830
+* ``ascatNgs`` PDF delivery (plots & statistics) #828
 
 [8.2.3]
 -------